+++ packages/cubejs-backend-shared/package.json
@@ -25,12 +25,16 @@
     "@cubejs-backend/linter": "^0.26.54",
     "@types/jest": "^26.0.16",
     "@types/node-fetch": "^2.5.7",
+    "@types/ramda": "^0.27.0",
+    "@types/sqlstring": "^2.3.0",
     "jest": "^26.6.3",
     "typescript": "~4.1.5"
   },
   "dependencies": {
     "env-var": "^6.3.0",
-    "node-fetch": "^2.6.1"
+    "node-fetch": "^2.6.1",
+    "ramda": "^0.27.0",
+    "sqlstring": "^2.3.0"
   },
   "publishConfig": {
     "access": "public"
