+++ packages/cubejs-backend-shared/src/index.ts
@@ -7,3 +7,4 @@ export * from './convert';
 export * from './helpers';
 export * from './machine-id';
 export * from './type-helpers';
+export * from './logger';
