+++ packages/cubejs-server-core/package.json
@@ -51,7 +51,6 @@
     "ramda": "^0.27.0",
     "semver": "^6.3.0",
     "serve-static": "^1.13.2",
-    "sqlstring": "^2.3.1",
     "uuid": "^3.3.3"
   },
   "devDependencies": {
